Temporary disable fMP4 remux

+++ TelegramUI/FetchVideoMediaResource.swift
@@ -145,7 +145,7 @@ public func fetchVideoLibraryMediaResource(postbox: Postbox, resource: VideoLibr
                             if stat(result.fileURL.path, &value) == 0 {
                                 if config.remuxToFMp4 {
                                     let tempFile = TempBox.shared.tempFile(fileName: "video.mp4")
-                                    if FFMpegRemuxer.remux(result.fileURL.path, to: tempFile.path) {
+                                    if false && FFMpegRemuxer.remux(result.fileURL.path, to: tempFile.path) {
                                         let _ = try? FileManager.default.removeItem(atPath: result.fileURL.path)
                                         subscriber.putNext(.moveTempFile(file: tempFile))
                                     } else {
@@ -230,7 +230,7 @@ func fetchLocalFileVideoMediaResource(postbox: Postbox, resource: LocalFileVideo
                     if stat(result.fileURL.path, &value) == 0 {
                         if config.remuxToFMp4 {
                             let tempFile = TempBox.shared.tempFile(fileName: "video.mp4")
-                            if FFMpegRemuxer.remux(result.fileURL.path, to: tempFile.path) {
+                            if false && FFMpegRemuxer.remux(result.fileURL.path, to: tempFile.path) {
                                 let _ = try? FileManager.default.removeItem(atPath: result.fileURL.path)
                                 subscriber.putNext(.moveTempFile(file: tempFile))
                             } else {
